port r52068 from trunk
git-svn-id: https://svn.r-project.org/R/branches/R-2-11-branch@52070 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/library/utils/R/example.R
@@ -36,6 +36,11 @@ function(topic, package = NULL, lib.loc = NULL, local = FALSE,
     tf <- tempfile("Rex")
     encoding <- "UTF-8"
     tools::Rd2ex(.getHelpFile(file), tf)
+    if (!file.exists(tf)) {
+        warning(gettextf("'%s' has a help file but no examples", topic),
+                domain = NA)
+        return(invisible())
+    }
     on.exit(unlink(tf))
     if(pkgname != "base")
         library(pkgname, lib.loc = lib, character.only = TRUE)
